refactor: break virtual keyboard into smaller modules
