+++ index.html
@@ -27,7 +27,7 @@
     <!-- This prevents FOUC for the virtual keyboard -->
     <link
       rel="preload"
-      href="./src/components/virtual-keyboard.css"
+      href="./src/ui/components/virtual-keyboard/styles.css"
       as="style"
     />
     <!-- Game styles -->
