+++ src/ui/components/virtual-keyboard/i18n.js
@@ -1,0 +1,6 @@
+export const defaultLanguage = "en"; // English
+
+export const LAYOUT = {
+  en: "qwertyuiop\nasdfghjkl\nBzxcvbnmE",
+  es: "qwertyuiop\nasdfghjklñ\nBzxcvbnmE",
+};
