+++ src/ui/components/virtual-keyboard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./virtual-keyboard.js";
